requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
pillow==10.1.0
xhtml2pdf==0.2.11
pandas==2.1.3
django-cors-headers==4.3.0
gunicorn==21.2.0
python-dotenv==1.0.0
drf-yasg==1.21.7  # API documentatio